feat(blast-db): round-2 hardening — composite signatures, cancel, CSWSH defence, audit
Closes second 20-item critique covering adjacent surfaces (shard, oracle,
NCBI catalogue resilience, WS security, frontend reactive bugs).

Backend:
- /shard daemon writes through ETag-aware _update_metadata so concurrent
  prepare-db/warmup writers can't clobber shard fields.
- /oracle refuses 409 when copy_status.phase in {partial, init_failed,
  copying}.
- check-updates precedence is composite_signature > signature_etag >
  source_version vs snapshot; multi-shard NCBI rotations now detected.
- New circuit breaker (api/routes/storage/common.py) trips on N consecutive
  403/5xx from NCBI and refuses calls for cooldown.
- New POST /api/storage/prepare-db/{db}/cancel — abort_copy on every
  pending blob + cancelled state.
- prepare_db lock registry caps at 256 with free-lock GC; poll loop
  honours a _SHUTDOWN_EVENT between batches.
- ncbi_catalogue: composite signature samples N=8 .tar.gz.md5 ETags +
  sha256-16 hex; explicit httpx Client timeout; full v5 suffix list.
- New api/services/db_ops_audit.py — JobState + history for prepare_db /
  shard / oracle / cancel actions.
- terminal_ws: CSWSH defence — refuse WebSocket upgrades whose Origin
  is neither same-origin nor in TERMINAL_WS_ALLOWED_ORIGINS.

Frontend:
- useBlastDb dedup ref so partial-copy toasts fire once per (db, phase);
  new handleCancel action.
- useDbPreviews memoised byName + skipNames suppresses preview HEADs for
  already-Ready DBs.
- BlastDbRow gains Cancel button during in-flight copy; Retry is per-DB
  (not gated by another DB's lock).

Tests: 4 new files (+28 cases) covering breaker, composite signatures,
409 on live daemon, cancel route, terminal origin allowlist, and ETag
fallback paths. 1200 backend tests pass.

+++ api/services/ncbi_catalogue.py
@@ -141,7 +141,7 @@ def _pick_signature_keys(db_name: str, keys: list[str]) -> list[str]:
             return md5s
         # Evenly-spaced sample including first and last.
         step = (len(md5s) - 1) / (n - 1)
-        sampled = [md5s[int(round(i * step))] for i in range(n)]
+        sampled = [md5s[round(i * step)] for i in range(n)]
         # Dedup preserving order.
         seen: set[str] = set()
         out: list[str] = []
